--- daemon/internal/grpc/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"victord/daemon/internal/grpc/server"
-)
-
-func main() {
-	server.StartGRPCServer()
-}
