+++ daemon/internal/index/service/interface.go
@@ -1,0 +1,11 @@
+package service
+
+import (
+	"context"
+	"victord/daemon/internal/dto"
+	"victord/daemon/internal/index/models"
+)
+
+type IndexService interface {
+	CreateIndex(ctx context.Context, idx *dto.CreateIndexRequest, name string) (*models.IndexResource, error)
+}
