+++ daemon/transport/http/routes/routes.go
@@ -1,0 +1,8 @@
+package routes
+
+const (
+	CreateIndexPath  = "/api/index/{indexName}"
+	InsertVectorPath = "/api/vector/{indexName}"
+	DeleteVectorPath = "/api/vector/{indexName}/{vectorID}"
+	SearchVectorPath = "/api/vector/{indexName}/search"
+)
